not used default config

+++ aria2/run.sh
@@ -1,8 +1,11 @@
 #/bin/sh
-
-curl -sSL -o /app/aria2.conf https://raw.githubusercontent.com/AnRightHand/rpi-dockers/master/aria2/aria2.conf
-mv -n /app/aria2.conf /config/aria2.conf
-touch /config/aria2.session
+if [[ ! -f "/config/aria2.conf" ]]; then
+  curl -f -sSL -o /app/aria2.conf https://raw.githubusercontent.com/AnRightHand/rpi-dockers/master/aria2/aria2.conf || true
+  mv -n /app/aria2.conf /config/aria2.conf
+fi
+if [[ ! -f "/config/aria2.session" ]]; then
+  touch /config/aria2.session
+fi
 
 /app/aria2c --conf-path=/config/aria2.conf
