+++ caddy/run.sh
@@ -1,7 +1,9 @@
 #/bin/sh
 
-mkdir -p /app 
-curl -sSL -o /app/Caddyfile https://raw.githubusercontent.com/AnRightHand/rpi-dockers/master/caddy/Caddyfile
-mv -n /app/Caddyfile /config/Caddyfile
+mkdir -p /app
+if [[ ! -f "/config/Caddyfile" ]]; then
+  curl -f -sSL -o /app/Caddyfile https://raw.githubusercontent.com/AnRightHand/rpi-dockers/master/caddy/Caddyfile || true
+  mv -n /app/Caddyfile /config/Caddyfile
+fi
 
 caddy -conf /config/Caddyfile
