+++ minidlna/run.sh
@@ -1,7 +1,9 @@
 #/bin/sh
 
 mkdir -p /app 
-curl -sSL -o /app/minidlna.conf https://raw.githubusercontent.com/AnRightHand/rpi-dockers/master/minidlna/minidlna.conf
-mv -n /app/minidlna.conf /config/minidlna.conf
+if [[ ! -f "/config/minidlna.conf" ]]; then
+  curl -sSL -o /app/minidlna.conf https://raw.githubusercontent.com/AnRightHand/rpi-dockers/master/minidlna/minidlna.conf || true
+  mv -n /app/minidlna.conf /config/minidlna.conf
+fi
   
 /usr/sbin/minidlnad -f /config/minidlna.conf -S
